authorization implemented

+++ src/api/songs/handler.js
@@ -27,6 +27,7 @@ class SongsHandler {
         genre,
         duration,
       } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
 
       const songId = await this._service.addSong({
         title,
@@ -34,6 +35,7 @@ class SongsHandler {
         performer,
         genre,
         duration,
+        owner: credentialId,
       });
 
       const response = h.response({
@@ -67,7 +69,8 @@ class SongsHandler {
   }
 
   async getSongsHandler() {
-    const songs = await this._service.getSongs();
+    const { id: credentialId } = request.auth.credentials;
+    const songs = await this._service.getSongs(credentialId);
     return {
       status: 'success',
       data: {
@@ -79,6 +82,9 @@ class SongsHandler {
   async getSongByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifySongOwner(id, credentialId);
       const song = await this._service.getSongById(id);
       return {
         status: 'success',
@@ -111,6 +117,10 @@ class SongsHandler {
     try {
       this._validator.validateSongPayload(request.payload);
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifySongOwner(id, credentialId);
+
       await this._service.editSongById(id, request.payload);
 
       return {
@@ -141,6 +151,9 @@ class SongsHandler {
   async deleteSongByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifySongOwner(id, credentialId);
       await this._service.deleteSongById(id);
       return {
         status: 'success',
